+++ src/Dosepy/app_model/app_model.py
@@ -1,7 +1,6 @@
 """Functions used as a model. VMC pattern."""
 
-from Dosepy.image import _is_RGB, _is_image_file, load, load_images, ImageLike, TiffImage
-from Dosepy.tools.files_to_image import equate_array_size, merge, stack_images
+from Dosepy.image import _is_RGB, _is_image_file, load, load_multiples, ImageLike, TiffImage
 import imageio.v3 as iio
 import numpy as np
 from numpy import ndarray
@@ -43,37 +42,11 @@ class Model:
             if iio.improps(file).shape != first_img_shape:
                 return False
         return True
-    
-    def load_calib_files(self, files: list) -> ImageLike:
-
-        if len(files) == 1:
-
-            return load(files[0], for_calib=True)
-        
-        elif len(files) > 1:
-
-            img = load(files[0], for_calib=True) # Placeholder
-            images = load_images(files)
-            equated_images = equate_array_size(images, axis=("width", "height"))
-            merged_images = merge(files, equated_images)
-            stacked = stack_images(merged_images, padding=6)
-            img.array = stacked.array
-
-            return img
 
 
     def load_files(self, files: list) -> ImageLike:
-        if len(files) == 1:
-            return load(files[0])
-        
-        else:
-            img = load(files[0]) # Placeholder
-            images = load_images(files)
-            equated_images = equate_array_size(images, axis=("width", "height"))
-            merged_images = merge(files, equated_images)
-            stacked = stack_images(merged_images, padding=6)
-            img.array = stacked.array
-            return img
+        # TODO Perform input validation
+        return load_multiples(files)
 
 
     def create_dosepy_lut(self, doses, roi):